requirements.txt
pyserial
PyQt6
pygame
numpy
openpyxl
pandas
pydirectinput
matplotlib
pytz